% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRaitRazor.R
\name{STRaitRazorLinux}
\alias{STRaitRazorLinux}
\title{STRaitRazor v3}
\usage{
STRaitRazorLinux(inputLocation, outputLocation, commandLineArguments = NULL,
  configFile = NULL, numberOfThreads = 4)
}
\arguments{
\item{inputLocation}{Path to fastq file.}

\item{outputLocation}{Path to the .txt returned by the STRait Razor v3 tool}

\item{commandLineArguments}{Any command line arguments passed to STRait Razor excluding -c and -p.}

\item{configFile}{The STRait Razor '.config' file to be used in the analysis (only ForenSeq and PowerSeq available).}

\item{numberOfThreads}{The number of threads used to run the analysis.}
}
\description{
Simple wrap function for calling the command line tool STRaitRazor v3 directly from R (on linux systems).
}
