% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRaitRazorLinux.R
\name{STRaitRazorLinuxSTRMPS}
\alias{STRaitRazorLinuxSTRMPS}
\title{STRait Razor output to STRMPS}
\usage{
STRaitRazorLinuxSTRMPS(outputLocation)
}
\arguments{
\item{outputLocation}{Location of the output created by the \link{STRaitRazorLinux}-function (or any run of the STRait Razor v3 command line tool).}
}
\value{
A list of tibbles used in the STRMPS package.
}
\description{
A function for converting the output created by STRait Razor to an object useable by the STRMPS-package.
}
