% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRaitRazor.R
\name{STRaitRazorSTRMPS.control}
\alias{STRaitRazorSTRMPS.control}
\title{STRaitRazor to STRMPS control function}
\usage{
STRaitRazorSTRMPS.control(commandLineArguments = NULL, configFile = NULL,
  numberOfThreads = 4)
}
\arguments{
\item{commandLineArguments}{Command line arguments passed to STRaitRazor v3.}

\item{configFile}{The STRaitRazor config file used for flanking regoin identification.}

\item{numberOfThreads}{The total number of cores used for parallelisation.}
}
\value{
A list containing default input.
}
\description{
Contains the arguments passed to the \link{STRaitRazor} function, when it is called through the \link{STRaitRazorSTRMPS} function.
}
