% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRaitRazor.R
\name{STRaitRazorSTRMPS}
\alias{STRaitRazorSTRMPS}
\title{STRaitRazor v3 output to STRMPS data structure.}
\usage{
STRaitRazorSTRMPS(inputLocation, control = STRaitRazorSTRMPS.control())
}
\arguments{
\item{inputLocation}{Path to fastq file.}

\item{control}{A control object containing additional input for the \link{STRaitRazor} function.}
}
\value{
A list of tibbles used by the STRMPS package.
}
\description{
A function for converting the STRaitRazor output to an object useable by the STRMPS-package.
}
