% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRaitRazor.R
\docType{class}
\name{stringCoverageList-class}
\alias{stringCoverageList-class}
\title{A string coverage list}
\description{
A list of tibbles, one for every marker, used to contain the sequencing information of STR MPS data.
}
